integration: split out innards of log/map tests
This allows for other test drivers.
